Update S3 processing service to catch CI imports in diff schema

+++ s3processingservice/imports3.py
@@ -27,9 +27,9 @@ def dump_message(message):
             bd = json.loads(message.body)
             for record in bd['Records']:
                 # Check it is for us
-                if record['eventName'] != 's3:ObjectCreated:Post' or record['s3']['bucket']['name'] != 'it_randd':
+                filename:str = record['s3']['object']['key']
+                if record['eventName'] != 's3:ObjectCreated:Post' or record['s3']['bucket']['name'] != 'it_randd' or not (filename.startswith('results/') or filename.startswith('ci_results/')):
                     next
-                filename = record['s3']['object']['key']
                 # Download new file from S3
                 print(f'Downloading file {filename}')
                 try:
@@ -40,21 +40,33 @@ def dump_message(message):
                 doc = res['Body'].read()
                 docconv = json.loads(doc)
                 res['Body'].close()
-                try:
-                    jsonschema.validate(instance = docconv, schema=schema)
-                    print(f'File {filename} has passed validation')
-                    # It has passed? Send it to postgres database
-                    # TODO: Catch exceptions from this and if it fails don't delete
-                    curs.execute("insert into returned_results (jsondata) values (%s)", (psycopg2.extras.Json(docconv),))
-                except jsonschema.exceptions.ValidationError as err:
-                    print(f'Downloaded file does not validate: {err.message}')
-                    pass
-                print(f'Deleting file {filename}')
-                # Delete it from S3
-                try:
-                    res_del = client.delete_object(Bucket='it_randd', Key=record['s3']['object']['key'])
-                except ClientError as ex:
-                    pass
+                if filename.startswith('results/'):
+                    try:
+                        jsonschema.validate(instance = docconv, schema=system_results_schema)
+                        print(f'File {filename} has passed validation')
+                        # It has passed? Send it to postgres database
+                        # TODO: Catch exceptions from this and if it fails don't delete
+                        curs.execute("insert into returned_results (jsondata) values (%s)", (psycopg2.extras.Json(docconv),))
+                    except jsonschema.exceptions.ValidationError as err:
+                        print(f'Downloaded file does not validate: {err.message}')
+                        pass
+                elif filename.startswith('ci_results/'):
+                    try:
+                        jsonschema.validate(instance = docconv, schema=ci_results_schema)
+                        print(f'File {filename} has passed validation')
+                        # It has passed? Send it to postgres database
+                        # TODO: Catch exceptions from this and if it fails don't delete
+                        curs.execute("insert into returned_results (jsondata) values (%s)", (psycopg2.extras.Json(docconv),))
+                    except jsonschema.exceptions.ValidationError as err:
+                        print(f'Downloaded file does not validate: {err.message}')
+                        pass
+                if args.delete:
+                    print(f'Deleting file {filename}')
+                    # Delete it from S3
+                    try:
+                        res_del = client.delete_object(Bucket='it_randd', Key=record['s3']['object']['key'])
+                    except ClientError as ex:
+                        pass
     pgconn.close()
     channel.basic_ack(message.delivery_tag)
 
@@ -68,12 +80,16 @@ parser.add_argument('password', metavar='password', type=str, nargs='?',
                     help='rabbitmq password', default='password')
 parser.add_argument('dsn', metavar='dsn', type=str, nargs='?',
                     help='postgres host', default='dbname=benchmarking user=postgres password=postgres')
+parser.add_argument('delete', metavar='delete', type=bool, nargs='?',
+                    help='Delete notified files', default=True)
 
 args = parser.parse_args()
 DSN = args.dsn
 
-with open('jsonschema.json', mode='r') as f:
-    schema = json.load(f)
+with open('system_results.json', mode='r') as f:
+    system_results_schema = json.load(f)
+with open('ci_results.json', mode='r') as f:
+    ci_results_schema = json.load(f)
 
 # Connect to Rabbit MQ server providing S3 notifications
 conn = Connection(host=args.host, userid=args.user,
